refactor: twitter constants
