+++ src/app/api/twitter/search/all/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from 'next/server.js';
 import { z } from 'zod';
 
-import { TWITTER_TIMELINE_OPTIONS } from '@/constants/index.js';
+import { TWITTER_TIMELINE_OPTIONS } from '@/constants/twitter.js';
 import { compose } from '@/helpers/compose.js';
 import { createSuccessResponseJSON } from '@/helpers/createResponseJSON.js';
 import { createAppOnlyTwitterClientV2 } from '@/helpers/createTwitterClientV2.js';
