+++ src/app/api/twitter/space/[spaceId]/route.ts
@@ -1,7 +1,7 @@
 import type { NextRequest } from 'next/server.js';
 
 import { MalformedError } from '@/constants/error.js';
-import { SPACE_OPTIONS } from '@/constants/index.js';
+import { SPACE_OPTIONS } from '@/constants/twitter.js';
 import { compose } from '@/helpers/compose.js';
 import { createSuccessResponseJSON } from '@/helpers/createResponseJSON.js';
 import { createAppOnlyTwitterClientV2 } from '@/helpers/createTwitterClientV2.js';
