+++ src/app/api/twitter/users/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from 'next/server.js';
 import { z } from 'zod';
 
-import { TWITTER_USER_OPTIONS } from '@/constants/index.js';
+import { TWITTER_USER_OPTIONS } from '@/constants/twitter.js';
 import { compose } from '@/helpers/compose.js';
 import { createSuccessResponseJSON } from '@/helpers/createResponseJSON.js';
 import { createTwitterClientV2 } from '@/helpers/createTwitterClientV2.js';
